TAN-35: move enterprise HTTP integration tests into tandem-enterprise-server
The enterprise HTTP tests lived in tandem-server (http::tests::enterprise)
but the routes they exercise are mounted only in tandem-enterprise-server
(via apply_routes), so they failed against tandem-server's base router and
were kept green only by a nextest `default-filter` exclusion. Move them to
the crate that owns the routes and run them against the real wiring.

tandem-server:
- Add a `test-support` feature exposing a public `test_support` module
  (test_state / next_event_of_type) so dependent crates can build a
  realistic AppState without leaking internals into the normal API.
- Move those helpers out of the cfg(test) test module into test_support
  and re-export them for the in-crate tests.
- Broaden the header-trusting local tenant resolver gate from cfg(test)
  to cfg(any(test, feature = "test-support")) so dependent-crate HTTP
  tests can drive per-tenant behavior (test-only; never in production).
- Delete the 5 orphaned routes_enterprise*.rs duplicates (dead: never in
  the module tree; the live copies are in tandem-enterprise-server).

tandem-enterprise-server:
- Add the moved suite as integration tests built on
  build_router_with_extensions(state, &[apply_routes]) — the same wiring
  serve_with_route_extensions uses in production. Split under the 2000-line
  CI cap into enterprise_http.rs + enterprise_http_google_drive.rs with
  shared body builders in tests/common.
- dev-deps: chrono, tandem-types, and tandem-server with the test-support
  feature.
- Two Google Drive import/reindex tests are #[ignore]d: they require a real
  MemoryManager (EmbeddingService::new), i.e. the local-embeddings stack,
  which is not available in the default test build. The other 24 enterprise
  tests now run in CI for the first time.

Drop the now-stale `http::tests::enterprise::` exclusion from the nextest
ci profile.
