+++ .config/nextest.toml
@@ -11,8 +11,7 @@ slow-timeout = { period = "60s", terminate-after = 5 }
 [profile.ci]
 default-filter = """
 not (
-  test(/^http::tests::enterprise::/)
-  | test(=app::state::automation::tests::code_workflow_with_connector_tool_allowlist_keeps_patch_tools)
+  test(=app::state::automation::tests::code_workflow_with_connector_tool_allowlist_keeps_patch_tools)
   | test(=app::state::tests::automations::automation_run_requeue_increments_attempt_counter)
   | test(=http::coder::issue_fix_handoff_tests::issue_fix_handoff_commits_and_pushes_worker_branch)
   | test(=http::tests::coder::coder_issue_fix_worker_uses_managed_worktree_for_git_repo)
